campus-connect/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
supabase==2.10.0
httpx==0.25.2
websockets==12.0
slowapi==0.1.9
python-dotenv==1.0.0
python-docx==1.1.0
